\name{corpora-package}
\alias{corpora-package}
\alias{corpora}
\docType{package}
\title{

corpora: Statistical Inference from Corpus Frequency Data

}
\description{

The \code{corpora} package provides a collection of functions for statistical inference
from corpus frequency data, as well as some convenience functions and example data sets.

It is a companion package to the open-source course \emph{Statistical Inference: a Gentle Introduction for Linguists and similar creatures} developed by Marco Baroni and Stefan Evert. Statistical methods implemented in the package are described and illustrated in the units of this course.


}
\details{

\bold{TODO:} overview of functions and data sets in package

}


\seealso{

\bold{TODO:} entry points into corpora documentation

}

\author{ 

Stefan Evert <\email{stefan.evert@fau.de}> 

}

\references{

The official homepage of the \code{corpora} package and the SIGIL course is \url{http://SIGIL.R-Forge.R-Project.org/}.

}

\keyword{ package }

\examples{

## TODO: basic usage examples?

}
