\name{corpora-package}
\alias{corpora-package}
\alias{corpora}
\docType{package}
\title{

corpora: statistical inference from corpus frequency data

}
\description{

The \code{corpora} provides some convenience functions and example data sets
for statistical inference from corpus frequency data.

It is a companion package for the SIGIL course
(\emph{Statistical Inference - a Gentle Introduction for Linguists and similar creatures})
developed by Marco Baroni and Stefan Evert.

}
\details{

\bold{ToDo:} overview of functions and data sets in package

}


\seealso{

  \bold{ToDo:} entry points into corpora documentation

}

\author{ Stefan Evert <\email{stefan.evert@uos.de}> }

\references{

  The official homepage of both the SIGIL course and the corpora package is
  \url{http://SIGIL.R-Forge.R-Project.org/}.

}

\keyword{ package }
%% ~~ optionally other keywords from "KEYWORDS.db", one per line ~~

\examples{

## TODO: basic usage examples?

}
