\name{corpora-package}
\alias{corpora-package}
\alias{corpora}
\docType{package}
\title{

corpora: Statistical Inference from Corpus Frequency Data

}
\description{

The \code{corpora} package provides a collection of functions for statistical inference
from corpus frequency data, as well as some convenience functions and example data sets.

}
\details{

\bold{ToDo:} overview of functions and data sets in package

}


\seealso{

\bold{ToDo:} entry points into corpora documentation

}

\author{ 

Stefan Evert <\email{stefan.evert@fau.de}> 

}

\references{

The official homepage of the \code{corpora} package is \url{http://SIGIL.R-Forge.R-Project.org/}.

}

\keyword{ package }

\examples{

## TODO: basic usage examples?

}
